automation: use go 1.9.2
Executing go test using 1.8.3 fails because of bad import resolution.
This behaiour is fixed in go 1.9.x and this patch installs go 1.9 inside
the mock running the various automation jobs.

note: as for now centos/rhel 7.5 doesn't have a package for go 1.9. This
may be fixed in 7.6 by the go-toolset package, but not sure.
