--- automation/defaults.conf
@@ -1,2 +1,0 @@
-ORG=github.com/ovirt
-REPO=ovirt-openshift-extensions
